services/inventory/requirements.txt
grpcio>=1.66.2
six==1.16.0
dapr==1.15.0
Flask>=3.1.1
